#152 address review feedback
